+++ include/mgdl-wii.h
@@ -22,8 +22,6 @@
 // muffintrap: Added new files in version 0.100.0-muffintrap
 #include "mgdl-wii/mgdl-gui.h"
 #include "mgdl-wii/mgdl-assert.h"
-#include "mgdl-wii/mgdl-defaultfont.h"
-#include "mgdl-wii/mgdl-input-wii.h"
 #include "mgdl-wii/mgdl-util.h"
 
 #endif // _GDL_WII_H