+++ include/mgdl-wii/mgdl-font.h
@@ -25,7 +25,7 @@ class FFont {
 	short		cw,ch;
 	void		*vList;
 	void		*tList;
-	short		firstIndex; // muffintrap: need to remember the first index to calculate offsets into vertex and uv arrays
+	char		firstIndex; // muffintrap: need to remember the first index to calculate offsets into vertex and uv arrays
 
 public:
 
@@ -38,24 +38,21 @@ public:
 	virtual	~FFont();
 
 
-	// muffintrap: modified to take less than 256 characters. This way the usual
+	// muffintrap: modified to be able to read in less than 256 characters. This way the usual
 	// spritefont images that have chacters between " !"#$% ... xyz{|}~" can be used
 	//! Binds the font handler class to an gdl::Image object of a font sheet.
 	/*!
 	 *	\details A font sheet consists of a simple array of 256 characters arranged in a 16x16 character grid. It is
 	 *	good enough for simple fonts (such as those found in retro inspired games) but will become problematic with
-	 *	dynamic fonts that have variable-width characters. If the sheet contains less characters or has
-	 *  different dimensions, the optional parameters for characters per row and
-	 *  first and last index can be specified.
+	 *	dynamic fonts that have variable-width characters. If the sheet does not start from ASCII code 0,
+	 *  the first character on the sheet can be specified
 	 *
 	 *	\param[in]	image	gdl::Image object of a font sheet.
 	 *	\param[in]	charw	Width of characters in pixels.
 	 *	\param[in]	charh	Height of characters in pixels.
-	 *	\param[in]	firstIndex	How many characters to skip  Default 0.
-	 *	\param[in]	lastIndex	Last character to be loaded. Exclusive!. Default 256.
-	 *	\param[in]	charactersPerRow	How many characters are on a row. Default 16.
+	 *	\param[in]	firstCharacter	The first character in the image
 	 */
-	void BindSheet(gdl::Image& image, short charw, short charh, short firstIndex = 0, short lastIndex = 256, short charactersPerRow = 16);
+	void BindSheet(gdl::Image& image, short charw, short charh, char firstCharacter);
 
 
 	//! Draw text.
@@ -88,120 +85,27 @@ public:
 	 */
 	void Printf(short x, short y, float scale, u32 col, const char* format, ... );
 
-
-};
-
-
-//! Font handling class
-/*!
- *	Unlike gdl::FFont, this font class supports variable-width characters but is currently limited to supporting
- *	DirectQB format fonts.
- */
-class Font {
-
-	char charWidth[256];
-	char charHeight;
-
-	void *vList;
-	void *tList;
-
-	gdl::Texture charSheet;
-
-public:
-
-
-	//! Class constructor.
-	Font();
-
-
-	//! Class deconstructor.
-	virtual ~Font();
-
-
-	//! Loads a font file.
-	/*!
-	 *	\details Currently, this function only supports DirectQB format font files. Support for TTF files may
-	 *	most likely never be implemented. Tools for converting TTF fonts to a sheet-style font format may
-	 *	most likely be implemented in the future.
-	 *
-	 *	\param[in]	*fileName	File name of the font file to load.
-	 *
-	 *	\returns Non-zero if the font file was successfully loaded, otherwise an error occurred.
-	 */
-	bool LoadFont(const char *fileName);
-
-
-	//! Loads a font from a memory buffer.
-	/*!
-	 *	\param[in]	*fontData	Pointer to font file data in memory.
-	 *
-	 *	\returns Non-zero if the font file was successfully loaded, otherwise an error occurred.
-	 */
-	 bool LoadFontMem(u_char *fontData);
-
-
-	//! Calculates the length of a string in pixels relative to the loaded font.
-	/*!
-	 *	\param[in]	*text	String to calculate the length of.
-	 *
-	 *	\returns Length of the string in pixels, zero if no font was loaded or the string is empty.
-	 */
-	int CalcStrLen(const char *text);
-
-	//! Overload alternative to CalcStrLen().
-	/*!
-	 *	\param[in]	*text	String to calculate the length of.
-	 *	\param[in]	len		Number of characters to calculate the length of.
-	 *
-	 *	\returns Length of the string in pixels, zero if no font was loaded or the string is blank.
-	 */
-	int CalcStrLen(const char *text, int len);
-
 	// muffintrap: added this to get the height of the font
 	//! Returns the height of this font in pixels.
 	/*!
 	 *
-	 *	\returns Height of the string in pixels, zero if no font was loaded or the string is blank.
+	 *	\returns Height of the font in pixels, zero if no font was loaded or the string is blank.
 	 */
-	int GetStrHeight();
+	short GetHeight();
 
-	//! Draw text.
+	// muffintrap: added this to get the width of the string in this font
+	//! Returns the width of this font in pixels.
 	/*!
-	 *	\details Draws a string of text using the loaded font. Because it only accepts a pre-made string of
-	 *	characters, it is not as flexible as printf() and gdl::Font::Printf()
 	 *
-	 *	\param[in]	*text	Text to draw.
-	 *	\param[in]	x		X coordinate of text (can be gdl::Centered to center text).
-	 *	\param[in]	y		Y coordinate of text.
-	 *	\param[in]	scale	Scale factor (1.0 is original size).
-	 *	\param[in]	col		Color of text (see gdl::Color::ColorValues or \ref colorMacros).
+	 *	\returns width of the string in pixels, zero if no font was loaded or the string is blank.
 	 */
-	void DrawText(const char *text, short x, short y, float scale, u_int col);
-
-
-	//! Draw text with a printf-like syntax.
-	/*!
-	 *	\details Similar to gdl::Font::DrawText() but is much more flexible as it accepts printf-style specifiers
-	 *	and formatting parameters.
-	 *
-	 *	\note The overall length of the string must be no longer than 256 characters.
-	 *
-	 *	\param[in]	x		X coordinate of text (can be gdl::Centered to center text).
-	 *	\param[in]	y		Y coordinate of text.
-	 *	\param[in]	scale	Scale factor (1.0 is original size).
-	 *	\param[in]	col		Color of text (see gdl::Color::ColorValues or \ref colorMacros).
-	 *	\param[in]	*format	Text to draw supporting printf specifiers and formatting parameters.
-	 *	\param[in]	...		Print arguments.
-	 */
-	void Printf(short x, short y, float scale, u32 col, const char* format, ... );
-
-
+	short GetWidth(const char* str);
 };
 
+// muffintrap: removed Font class from version 0.100.0-muffintrap
+// since I don't think DirectQB fonts get much use anymore.
+
 /*!	@} */
-
-
 }
-
 
 #endif // _GDL_WII_FONT_H